pyproject.toml
[project]
name = "mcp-postgres-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]
